Process SRV record before any other record
It is possible that a A record or a AAAA record comes before
a SRV record in the answer section of a mDNS response. When
a A record or a AAAA record is being proccessed, the
corresponding ServiceEntry is not yet available. By the time
the SRV record is being proccessed, the ServiceEntry would
contain no IPv4 address and IPv6 address, though records are
available in the DNS answer section.